requirements.txt
tweepy>=4.14.0
pandas>=1.5.0
transformers>=4.30.0
torch>=2.0.0
spacy>=3.5.0
nltk>=3.8.0
streamlit>=1.22.0
emoji>=2.2.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
scikit-learn>=1.2.0
plotly>=5.13.0